calculate_savings_variable_type

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 #define variable and accept user input
-monthly_income = int(input("Enter your monthly income:"))
-monthly_expenses = int(input("Enter your total monthly expenses:"))
+monthly_income = float(input("Enter your monthly income:"))
+monthly_expenses = float(input("Enter your total monthly expenses:"))
 
 
 #calculate monthly savings
